split off I/O functions
